fix httpstatus response

+++ src/main/java/com/KhoiCG/TMDT/orderService/dto/CouponCheckRequest.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CouponCheckRequest {
     private String code;
-    private Double orderAmount;
+    private Long orderAmount;
 }
